+++ TUM Campus App/TumSexyManager.swift
@@ -7,13 +7,29 @@
 //
 
 import Sweeft
+import Fuzzi
 
-final class TumSexyManager: MemoryCachedManager {
+struct SexyIndex {
+    let entries: [SexyEntry]
+    let tree: SearchTree<SexyEntry>?
+}
+
+final class TumSexyManager: MemoryCachedManager, SearchManager {
     
     typealias DataType = SexyEntry
     
     var config: Config
-    var cache: Cache<[SexyEntry]>?
+    var tree: SearchTree<SexyEntry>?
+    var indexCache: Cache<SexyIndex>?
+    
+    var cache: Cache<[SexyEntry]>? {
+        get {
+            return indexCache.map { $0.map { $0.entries } }
+        }
+        set {
+            indexCache = newValue.map { $0.map { .init(entries: $0, tree: $0.searchTree()) } }
+        }
+    }
     
     var requiresLogin: Bool {
         return false
@@ -23,10 +39,25 @@ final class TumSexyManager: MemoryCachedManager {
         return .time(.aboutOneWeek)
     }
     
+    var categoryKey: SearchResultKey {
+        return .sexy
+    }
+    
     init(config: Config) {
         self.config = config
     }
-    
+
+    func search(query: String) -> Promise<[SexyEntry], APIError> {
+        guard let tree = defaultMaxCache.validValue(in: indexCache)?.tree else {
+            return fetch().flatMap { _ in
+                return self.search(query: query)
+            }
+        }
+        return async(runQueue: .global()) {
+            return tree.search(query: query)
+        }
+    }
+
     func performRequest(maxCache: CacheTime) -> Response<[SexyEntry]> {
         return config.tumSexy.doJSONRequest(to: .sexy,
                                             maxCacheTime: maxCache).map { (json: JSON) in